requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
openai
PyJWT==2.8.0
tqdm